refactor: imports, switch cases, and hook loading state as commented
