+++ package.json
@@ -62,8 +62,6 @@
     "mobx-react-lite": "^3.4.0",
     "next": "^13.0.0",
     "next-seo": "^5.8.0",
-    "osmojs": "^14.0.0-rc.0",
-    "protobufjs": "^7.2.2",
     "react": "^18.2.0",
     "react-ace": "^10.1.0",
     "react-dom": "^18.2.0",
