+++ src/lib/components/copy/CopyTemplate.tsx
@@ -1,4 +1,3 @@
-import type { TooltipProps } from "@chakra-ui/react";
 import { Box, Tooltip, useClipboard } from "@chakra-ui/react";
 import { useEffect } from "react";
 
@@ -6,14 +5,12 @@ interface CopyTemplateProps {
   value: string;
   copyLabel?: string;
   triggerElement: JSX.Element;
-  tooltipBgColor?: TooltipProps["bgColor"];
 }
 
 export const CopyTemplate = ({
   value,
   copyLabel = "Copied!",
   triggerElement,
-  tooltipBgColor = "honeydew.darker",
 }: CopyTemplateProps) => {
   const { onCopy, hasCopied, setValue } = useClipboard(value);
   useEffect(() => setValue(value), [value, setValue]);
@@ -25,7 +22,7 @@ export const CopyTemplate = ({
       label={copyLabel}
       placement="top"
       arrowSize={8}
-      bgColor={tooltipBgColor}
+      bgColor="honeydew.darker"
     >
       <Box
         onClick={(e) => {
