--- src/lib/pages/tx-details/components/tx-message/msg-receipts/mapping.ts
@@ -1,7 +1,0 @@
-export const voteOption = {
-  VOTE_OPTION_UNSPECIFIED: "Empty",
-  VOTE_OPTION_YES: "Yes",
-  VOTE_OPTION_NO: "No",
-  VOTE_OPTION_NO_WITH_VETO: "NoWithVeto",
-  VOTE_OPTION_ABSTAIN: "Abstain",
-};
