+++ src/lib/types/currency/balance.ts
@@ -3,9 +3,9 @@ import type { Option } from "lib/types";
 export interface Balance {
   amount: string;
   id: string;
-  name?: Option<string>;
+  name?: string;
   precision: number;
-  symbol?: Option<string>;
+  symbol?: string;
   type?: string;
   price?: number;
 }
@@ -21,7 +21,7 @@ export interface AssetInfo {
   slugs: string[];
   symbol: string;
   type: string;
-  price: number;
+  price: Option<number>;
 }
 
 export interface BalanceWithAssetInfo {
