+++ src/lib/utils/tx/extractTxDetails.ts
@@ -1,17 +1,12 @@
 import type { Log } from "@cosmjs/stargate/build/logs";
 import { snakeCase } from "snake-case";
 
-import type { MsgBody, TypeUrl } from "lib/services/tx";
+import type { TypeUrl } from "lib/data";
+import type { MsgBody } from "lib/services/tx";
 import type { Option } from "lib/types";
 
 import { findAttr } from "./findAttr";
-import type {
-  MsgReturnType,
-  MsgStoreCodeDetails,
-  MsgInstantiateDetails,
-  MsgInstantiate2Details,
-  MsgSubmitProposalDetails,
-} from "./types";
+import type { MsgReturnType } from "./types";
 
 type MsgBodyWithoutType = Omit<MsgBody, "@type">;
 
@@ -41,28 +36,20 @@ export const extractTxDetails = <T extends TypeUrl>(
     case "/cosmwasm.wasm.v1.MsgStoreCode":
       return {
         type,
-        ...(msgBody as Omit<MsgStoreCodeDetails, "code_id">),
+        ...msgBody,
         code_id: findAttr(log, "store_code", "code_id"),
       } as MsgReturnType<T>;
     case "/cosmwasm.wasm.v1.MsgInstantiateContract":
-      return {
-        type,
-        ...(msgBody as Omit<MsgInstantiateDetails, "contract_address">),
-        contract_address: findAttr(log, "instantiate", "_contract_address"),
-      } as MsgReturnType<T>;
     case "/cosmwasm.wasm.v1.MsgInstantiateContract2":
       return {
         type,
-        ...(msgBody as Omit<MsgInstantiate2Details, "contract_address">),
+        ...msgBody,
         contract_address: findAttr(log, "instantiate", "_contract_address"),
       } as MsgReturnType<T>;
     case "/cosmos.gov.v1beta1.MsgSubmitProposal":
       return {
         type,
-        ...(msgBody as Omit<
-          MsgSubmitProposalDetails,
-          "proposal_id" | "proposal_type"
-        >),
+        ...msgBody,
         proposal_id: findAttr(log, "submit_proposal", "proposal_id"),
         proposal_type: findAttr(log, "submit_proposal", "proposal_type"),
       } as MsgReturnType<T>;
