+++ src/lib/utils/tx/findAttr.ts
@@ -1,16 +1,15 @@
-import type { Log } from "@cosmjs/stargate/build/logs";
-import { findAttribute } from "@cosmjs/stargate/build/logs";
+import { logs } from "@cosmjs/stargate";
 
 import type { Option } from "lib/types";
 
 export const findAttr = (
-  log: Option<Log>,
+  log: Option<logs.Log>,
   eventType: string,
   attrKey: string
 ): Option<string> => {
   if (!log) return undefined;
   try {
-    return findAttribute([log], eventType, attrKey).value;
+    return logs.findAttribute([log], eventType, attrKey).value;
   } catch {
     return undefined;
   }
